after a lot of checkups

+++ CPP Module 09/ex01/RPN.hpp
@@ -22,7 +22,7 @@ class RPN
 		};
 	private:
 		std::string m_arg;
-		std::stack<int> rpn;
+		std::stack<double> rpn;
 
 };
